+++ src/Hate/Graphics/Types.hs
@@ -6,14 +6,6 @@ import qualified Graphics.Rendering.OpenGL as GL
 import Data.Vect.Float
 import Data.Vect.Float.Instances()
 
--- supposedly needs more vertex streams and pipelines in the future
-data GraphicsState = GraphicsState { 
-    solidColorPipeline :: Pipeline,
-    texturingPipeline :: Pipeline,
-    globalVertexStream :: VertexStream,
-    screenSize :: (Int, Int)
-}
-
 -- |A general type for a graphical mesh, either in indexed or raw form.
 data VertexStream = VertexStream {
     vao :: GL.VertexArrayObject,
